+++ lnb-server/src/assistant.rs
