+++ lnb-server/src/impls/function/get_illust_url.rs
