+++ lnb-server/src/impls/function/image_generator.rs
