+++ lnb-server/src/impls/function/local_info.rs
