+++ lnb-server/src/impls/function/self_info.rs
