+++ lnb-server/src/impls/llm/openai/chat_completion.rs
