+++ lnb-server/src/impls/llm/openai/responses.rs
