+++ lnb-server/src/model.rs
