+++ lnb-server/src/text.rs
